+++ changelog.md
@@ -7,7 +7,8 @@ This version is not released yet and is under active development.
 ```
 
 - \[mpm\] Allow global `upgrade` of a subset of packages from the command line.
-- \[mpm\] Add a `-A/--all` option to `upgrade` operation.
+- \[mpm\] Add a `-A`/`--all` option to `upgrade` operation.
+- \[mpm\] add a `-d`/`--duplicates` option to `installed` operation to only show packages sharing the same ID accross multiple managers.
 - \[mpm\] Alias `list` operation to `installed`, `uninstall` to `remove`, `update` to `upgrade` and `lock`/`freeze` to `backup`.
 - \[pipx\] Implement `outdated` operation.
 - \[mpm\] Rename all `*-like` labels to `*-based` to help finer identification of families.